+++ src/main/java/ru/forinnyy/tm/command/project/ProjectRemoveByIndexCommand.java
@@ -1,7 +1,9 @@
 package ru.forinnyy.tm.command.project;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import ru.forinnyy.tm.exception.entity.AbstractEntityException;
+import ru.forinnyy.tm.exception.entity.ProjectNotFoundException;
 import ru.forinnyy.tm.exception.field.AbstractFieldException;
 import ru.forinnyy.tm.exception.user.AbstractUserException;
 import ru.forinnyy.tm.model.Project;
@@ -9,17 +11,21 @@ import ru.forinnyy.tm.util.TerminalUtil;
 
 public final class ProjectRemoveByIndexCommand extends AbstractProjectCommand {
 
+    @NotNull
     private static final String NAME = "project-remove-by-index";
 
+    @NotNull
     private static final String DESCRIPTION = "Remove project by index.";
 
+    @NotNull
     @Override
-    public @NotNull String getDescription() {
+    public String getDescription() {
         return DESCRIPTION;
     }
 
+    @NotNull
     @Override
-    public @NotNull String getName() {
+    public String getName() {
         return NAME;
     }
 
@@ -27,9 +33,10 @@ public final class ProjectRemoveByIndexCommand extends AbstractProjectCommand {
     public void execute() throws AbstractEntityException, AbstractFieldException, AbstractUserException {
         System.out.println("[REMOVE PROJECT BY INDEX]");
         System.out.println("ENTER INDEX:");
-        final Integer index = TerminalUtil.nextNumber() -1;
-        final Project project = getProjectService().removeByIndex(index);
-        final String userId = getUserId();
+        @NotNull final Integer index = TerminalUtil.nextNumber() -1;
+        @Nullable final Project project = getProjectService().removeByIndex(index);
+        if (project == null) throw new ProjectNotFoundException();
+        @NotNull final String userId = getUserId();
         getProjectTaskService().removeProjectById(userId, project.getId());
     }
